Add scheduing options for workouts, first version, #8
The schedules allow to group workouts. If the schedule is marked as a loop,
the workouts are looped till the schedule is deactivated by the user. In the
dashboard the 'correct' workout is shown, i.e. the active one from a schedule
or if there are none, simply the last one.