Fix major bug on offcanvas menu on android
It seems the default browser/rendering engine used in some versions of
android has problems with the CSS used for the offcanvas menu. This fix
is really a stopgag meassure since it makes it usable but somewhat changes
the behaviour. Ideally there should be better solution (there are also some
weird effects when swiping the menu sidewides).