Remove user Foreign Key from ingredients
This field was removed to make the author field a string only, this is
needed when dumping the ingredient database, as the referenced users do
not exist in the other instances.

Also, the list of possible submission status was reduced to three, just
as the exercises. The others were not really used.